fixed lines
